+++ src/body.rs
@@ -254,11 +254,11 @@ impl AsRef<[u8]> for Binary {
     }
 }
 
-impl Responder for Binary {
+impl<S> Responder<S> for Binary {
     type Item = HttpResponse;
     type Error = Error;
 
-    fn respond_to<S>(self, req: &HttpRequest<S>) -> Result<HttpResponse, Error> {
+    fn respond_to(self, req: &HttpRequest<S>) -> Result<HttpResponse, Error> {
         Ok(HttpResponse::build_from(req)
             .content_type("application/octet-stream")
             .body(self))