+++ src/json.rs
@@ -116,11 +116,11 @@ where
     }
 }
 
-impl<T: Serialize> Responder for Json<T> {
+impl<S, T: Serialize> Responder<S> for Json<T> {
     type Item = HttpResponse;
     type Error = Error;
 
-    fn respond_to<S>(self, req: &HttpRequest<S>) -> Result<HttpResponse, Error> {
+    fn respond_to(self, req: &HttpRequest<S>) -> Result<HttpResponse, Error> {
         let body = serde_json::to_string(&self.0)?;
 
         Ok(req