+++ src/lib.rs
@@ -5,7 +5,7 @@
 //! use actix_web::{server, App, Path, Responder};
 //! # use std::thread;
 //!
-//! fn index(info: Path<(String, u32)>) -> impl Responder {
+//! fn index(info: Path<(String, u32)>) -> impl Responder<()> {
 //!     format!("Hello {}! id:{}", info.0, info.1)
 //! }
 //!